Add poster_path to Movie model
The TMDB api assigns to every movie a relative URL that is used for retrieving the poster image. (It can be found on the “poster_path” key of every movie entity returned by the api) Better to cache it on the Movie model so it saves tons of HTTP requests when displaying the user’s wishlist, for instance.